+++ model/src/main/scala/by/sideproject/videocaster/model/filestorage/FileMeta.scala
@@ -2,4 +2,4 @@ package by.sideproject.videocaster.model.filestorage
 
 import by.sideproject.videocaster.model.BaseObject
 
-case class FileMeta(id: Option[Long], downloadURL: String, secondaryDownloadURL: Option[String], name: String, placement: String, path: String) extends BaseObject[Long]
+case class FileMeta(id: Option[Int], downloadURL: String, secondaryDownloadURL: Option[String], name: String, placement: String, path: String) extends BaseObject[Int]
