+++ services/storage/base/src/main/scala/by/sideproject/videocaster/services/storage/base/dao/VideoItemDetailsDAO.scala
@@ -2,4 +2,4 @@ package by.sideproject.videocaster.services.storage.base.dao
 
 import by.sideproject.videocaster.model.VideoItemDetails
 
-trait VideoItemDetailsDAO extends BaseDAO[VideoItemDetails, Long]
+trait VideoItemDetailsDAO extends BaseDAO[VideoItemDetails, Int]
